+++ crates/trestle/src/type_check/typed_ast.rs
@@ -55,7 +55,8 @@ pub struct TypeCheckedExpression {
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum TypeCheckedLiteral {
-    Int(usize),
+    /// See [`parse::ast::Literal::Int`](crate::parse::ast::Literal::Int) for why this is `i64`.
+    Int(i64),
     Bool(bool),
     Float(f64),
     String(String),
